Experiment with proptest.
Found a problem - numbers can be i64, but instance Ids are u64.

+++ polar/src/types.rs
@@ -229,7 +229,7 @@ pub struct Operation {
 
 #[derive(Debug, Clone, Serialize, Deserialize, Eq, PartialEq, Hash)]
 pub enum Value {
-    Integer(i64),
+    Integer(u64),
     String(String),
     Boolean(bool),
     ExternalInstance(ExternalInstance),
